// types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
      certificates: {
        Row: {
          category: string
          created_at: string
          credential_url: string | null
          display_order: number
          id: string
          image_url: string | null
          issue_date: string
          issuer: string
          title_en: string
          title_es: string
        }
        Insert: {
          category: string
          created_at?: string
          credential_url?: string | null
          display_order?: number
          id?: string
          image_url?: string | null
          issue_date: string
          issuer: string
          title_en: string
          title_es: string
        }
        Update: {
          category?: string
          created_at?: string
          credential_url?: string | null
          display_order?: number
          id?: string
          image_url?: string | null
          issue_date?: string
          issuer?: string
          title_en?: string
          title_es?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}